Fixed file folder structure. Added a method to get all games in the catalog. Added three tests and ensured pom.xml file has necessary dependency for JUnit.
